+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 ## CHANGELOG
 
+### v0.36.3
+
+* Enhancement: Add new files_modified condition #741
+* Enhancement: Enable to set default toolchain for all core cargo tasks #743
+* Documentation: Add section on default task definition #745
+
 ### v0.36.2 (2022-10-18)
 
 * Enhancement: Add env var support in env_files values #723